Vanilla 12 10 (#52)
* CI: Periodic Controller Database Update (#3226)

Updating Controller Database

Co-authored-by: OpenGOALBot <OpenGOALBot@users.noreply.github.com>

* [Loader] Unload less stuff per frame (#3227)

This should avoid the stuttering due to slow unloading on some drivers.

I also turned up the amount of stuff we load per frame since nobody has
been complaining about stutters there, but there has been a few cases of
levels loading in too slowly.

(this only changes graphics, not actual GOAL level load times).

* [jak2] adjust progress generic item sizes + add scrolling effect (#3231)

Makes most of the options larger to reduce the amount of dead space and
adds a smooth scrolling effect instead of having "pages".


![imagem](https://github.com/open-goal/jak-project/assets/7569514/fc483dea-226c-44ff-96e0-680e89a6df81)

Button remap menu has no scrolling. Fixes #2984

* [jak2] split language and text language (#3233)

* [jak2] Support per-proto shrub visibility (#3228)

Fixes the issue reported in
https://github.com/open-goal/jak-project/issues/3168,
https://github.com/open-goal/jak-project/issues/3189,
https://github.com/open-goal/jak-project/issues/3166,
https://github.com/open-goal/jak-project/issues/3152,
https://github.com/open-goal/jak-project/issues/3224 where a small part
of the pipes in the `strip-grenade` mission appear, but shouldn't.

* Allow replacing all instances of a texture with one texture (#3234)

There are art groups that are present in multiple levels, and that means
that also their textures are present in multiple levels. With texture
replacement, currently we need to make replacements for all instances if
we want it replaced everywhere, but this is not ideal, especially when
you make changes to your replacement texture and now you have to put it
in each folder again.

I added a way to replace all instances of a texture, by letting
texture-replacer people put their replacements into an '_all' folder. I
set up the logic in such a way that if you have a replacement for the
texture in its corresponding folder, it will take priority over a
replacement that you placed into the '_all' folder.

I personally found this very useful for replacing guard textures. The
guards appear in a lot of levels. But ideally you want them to look the
same everywhere. And that is why I looked into this and made a PR.

Oh and I changed what is printed in the 'Replacing ' part because it was
printing the path to our replacement, which didn't look nicely when
several textures got replaced by the same replacement from the '_all'
folder. So now it will print the original texture's page and name, I
think this information is more useful anyway.

* jak2: add vehicle hijack lines PC cheat (#3205)

Co-authored-by: Tyler Wilding <xtvaser@gmail.com>

* [jak2] Fix "Select controller" (#3191)

Co-authored-by: Martino Fontana <tinozzo123@gmail.com>

* [jak2] Support STICK_DEADZONE (#3239)

fixes #3235

* [jak2] add resolutions menu (#3238)

* [jak2] fix use-after-free bug in nav enemies (#3240)

Fixes #3153

* [jak2] fix health bar color fading (#3241)

Fixes #3237

* [jak2] add game territory menu option (#3242)

* save & load game territory setting (#3243)

* jak2: 90 fps menu option (#3245)

Closes #3223

* 989snd: sound bank loading overhaul (#3185)

* [jak2] Fix unicode issue with GAME.CGO load (#3254)

Fix for https://github.com/open-goal/jak-project/issues/3244

* merge vanilla 12-10

---------

Co-authored-by: OpenGOAL Bot <99294829+OpenGOALBot@users.noreply.github.com>
Co-authored-by: OpenGOALBot <OpenGOALBot@users.noreply.github.com>
Co-authored-by: water111 <48171810+water111@users.noreply.github.com>
Co-authored-by: ManDude <7569514+ManDude@users.noreply.github.com>
Co-authored-by: Luminar Light <18116946+LuminarLight@users.noreply.github.com>
Co-authored-by: Hat Kid <6624576+Hat-Kid@users.noreply.github.com>
Co-authored-by: Tyler Wilding <xtvaser@gmail.com>
Co-authored-by: SuperSamus <40663462+SuperSamus@users.noreply.github.com>
Co-authored-by: Martino Fontana <tinozzo123@gmail.com>
Co-authored-by: Brent Hickey <brent.hickey@icloud.com>
Co-authored-by: Ziemas <ziemas@ziemas.se>